+++ lib/base/index.dart
@@ -8,4 +8,4 @@ export 'models/iresponse_model.dart';
 
 export 'state/base_state.dart';
 
-export 'view/base_view.dart';
+export 'view/base_view_mvvm.dart';