+++ lib/base/view/base_view_mvvm_with_mobx.dart
@@ -1,0 +1,12 @@
+import 'package:mobx/mobx.dart';
+
+import '../index.dart';
+
+class BaseViewMVVMWithMobx<T extends Store> extends BaseViewMVVM<T> {
+  const BaseViewMVVMWithMobx({
+    super.key,
+    required super.viewModel,
+    required super.onPageBuilder,
+    required super.onModelReady,
+  });
+}