+++ CHANGELOG.md
@@ -9,3 +9,14 @@
 ## 0.0.3
 
 - Update README.md
+
+## 0.0.4
+
+- BaseView replaced with BaseViewMVVMWithMobx and extended with BaseViewMVVM.
+- BaseViewMVVMWithMobx is used with mobx.
+- New a BaseView added.
+- BaseViewMVVM added.
+- NotFoundPage edited.
+  - appBar (optional)
+  - goBack (optional)
+  - goBackWidget (optional)